ndt7: move download's sender in a separate file
Continue reorganizing the download as a pipeline.
