+++ ndt7/server/download/sender/sender.go
@@ -1,0 +1,66 @@
+// Package sender implements the download sender.
+package sender
+
+import (
+	"math/rand"
+	"time"
+
+	"github.com/gorilla/websocket"
+	"github.com/m-lab/ndt-server/logging"
+	"github.com/m-lab/ndt-server/ndt7/model"
+	"github.com/m-lab/ndt-server/ndt7/spec"
+)
+
+// defaultTimeout is the default value of the I/O timeout.
+const defaultTimeout = 7 * time.Second
+
+// makePreparedMessage generates a prepared message that should be sent
+// over the network for generating network load.
+func makePreparedMessage(size int) (*websocket.PreparedMessage, error) {
+	const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+	data := make([]byte, size)
+	// This is not the fastest algorithm to generate a random string, yet it
+	// is most likely good enough for our purposes. See [1] for a comprehensive
+	// discussion regarding how to generate a random string in Golang.
+	//
+	// .. [1] https://stackoverflow.com/a/31832326/4354461
+	for i := range data {
+		data[i] = letterBytes[rand.Intn(len(letterBytes))]
+	}
+	return websocket.NewPreparedMessage(websocket.BinaryMessage, data)
+}
+
+// Start is a reducer that drains the measurements while sending messages
+// mixed with measurements over the websocket connection.
+func Start(conn *websocket.Conn, measurements <-chan model.Measurement) error {
+	defer func() {
+		for range measurements {
+			// make sure we drain the channel
+		}
+	}()
+	logging.Logger.Debug("Generating random buffer")
+	const bulkMessageSize = 1 << 13
+	preparedMessage, err := makePreparedMessage(bulkMessageSize)
+	if err != nil {
+		return err
+	}
+	logging.Logger.Debug("Start sending data to client")
+	conn.SetReadLimit(spec.MinMaxMessageSize)
+	for {
+		select {
+		case m, ok := <-measurements:
+			if !ok {
+				return nil // the measurer told us it's time to stop running
+			}
+			conn.SetWriteDeadline(time.Now().Add(defaultTimeout))
+			if err := conn.WriteJSON(m); err != nil {
+				return err
+			}
+		default:
+			conn.SetWriteDeadline(time.Now().Add(defaultTimeout))
+			if err := conn.WritePreparedMessage(preparedMessage); err != nil {
+				return err
+			}
+		}
+	}
+}
